Move DeepStats to separate file as dirnode is a bit too large to work with comfortably.
